+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Option to register http method in route callbacks in `Server.registerRouteCallback`.
 - Route callback can be registered as file in `rsconfig.json`.
 - Using promises in route callbacks.
+- Client error handling.
 ### Breaking changes
 - Removed `Button` component.
 - Removed `res.render` function.